refactor(streamer): enforce capture/encode/packetize/transport/pipeline srp tree
