+++ host/rust/crates/wbeamd-streamer/src/capture/backend/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod wayland;
+pub(crate) mod x11;
