+++ host/rust/crates/wbeamd-streamer/src/capture/backend/wayland.rs
