+++ host/rust/crates/wbeamd-streamer/src/encode/h264.rs
