+++ host/rust/crates/wbeamd-streamer/src/encode/rawpng.rs
