+++ host/rust/crates/wbeamd-streamer/src/encode/selector.rs
