+++ host/rust/crates/wbeamd-streamer/src/main.rs
@@ -1,7 +1,7 @@
-mod backend;
 mod capture;
 mod cli;
-mod encoder;
+mod encode;
+mod packetize;
 mod pipeline;
 mod transport;
 
@@ -16,9 +16,9 @@ use gst::glib;
 use gst::prelude::*;
 use gstreamer as gst;
 
-use backend::prepare_capture;
+use capture::prepare_capture;
 use cli::{resolve_profile, Args};
-use encoder::{is_hevc, is_png};
+use encode::{is_hevc, is_png};
 use pipeline::make_pipeline;
 use transport::{hello_mode_bits, spawn_sender, HELLO_CODEC_HEVC, HELLO_CODEC_PNG};
