+++ host/rust/crates/wbeamd-streamer/src/pipeline/mod.rs
@@ -1,0 +1,4 @@
+mod builder;
+mod profile;
+
+pub(crate) use builder::make_pipeline;
